@@ -14,17 +14,17 @@
     "create-new-episode": "node -r esbuild-register scripts/create-new-episode.ts"
   },
   "dependencies": {
-    "@devtools-ds/icon": "^1.1.0",
-    "@devtools-ds/navigation": "^1.1.0",
-    "@devtools-ds/object-inspector": "^1.1.0",
-    "@devtools-ds/themes": "^1.1.0",
+    "@devtools-ds/icon": "^1.2.1",
+    "@devtools-ds/navigation": "^1.2.1",
+    "@devtools-ds/object-inspector": "^1.2.1",
+    "@devtools-ds/themes": "^1.2.1",
     "@radix-ui/react-label": "^2.0.2",
     "@vercel/analytics": "^1.0.1",
     "clsx": "^1.1.1",
     "googleapis": "^122.0.0",
     "gray-matter": "^4.0.3",
-    "next": "^13.4.12",
-    "next-mdx-remote": "^2.1.4",
+    "next": "^14.0.0",
+    "next-mdx-remote": "^4.4.1",
     "query-string": "^7.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -33,7 +33,7 @@
     "use-query-params": "^1.2.2"
   },
   "devDependencies": {
-    "@design-systems/next-esm-css": "^2.17.2",
+    "@design-systems/next-esm-css": "^2.18.0",
     "@types/react": "^17.0.3",
     "@types/rss": "^0.0.31",
     "autoprefixer": "^10.2.5",
